+++ multi-token-edit/scripts/presets/containerApp.js
@@ -9,7 +9,7 @@ import { PresetConfig } from './editApp.js';
 import { PresetBrowser } from './browser/browserApp.js';
 import { Preset } from './preset.js';
 import { Spawner } from './spawner.js';
-import { FolderState, isVideo } from './utils.js';
+import { exportPresets, FolderState, isVideo } from './utils.js';
 import { FileIndexer, IndexerForm } from './fileIndexer.js';
 
 export async function registerPresetHandlebarPartials() {
@@ -487,7 +487,8 @@ export class PresetContainer extends FormApplication {
   async _onImportFauxScene(item) {
     const preset = await PresetAPI.getPreset({ uuid: item.data('uuid') });
     const scene = await fromUuid(preset.data[0].uuid);
-    return game.scenes.importFromCompendium(scene.compendium, scene.id, {}, { renderSheet: true });
+    if (scene) game.scenes.importFromCompendium(scene.compendium, scene.id, {}, { renderSheet: true });
+    else ui.notifications.warn('Failed to find scene with UUID: ' + preset.data[0].uuid);
   }
 
   async _onSpawnScene(item) {
@@ -856,6 +857,11 @@ export class PresetContainer extends FormApplication {
 
   async _onFolderDelete(uuid, { render = true, deleteAll = false } = {}) {
     throw new Error('A subclass of the PresetContainer must implement the _onFolderDelete method.');
+  }
+
+  async _onExportSelectedPresets() {
+    const [selected, _] = await this._getSelectedPresets();
+    exportPresets(selected);
   }
 
   async _onExportSelectedPresetsToComp() {